+++ models/qos.go
@@ -20,21 +20,21 @@ type Qos struct {
 	BatchTransactionsPerSecondLimit    float64              `json:"batch_transactions_per_second_limit"`
 	LimitingDataLagStorageServer       Lag                  `json:"limiting_data_lag_storage_server"`
 	LimitingDurabilityLagStorageServer Lag                  `json:"limiting_durability_lag_storage_server"`
-	LimitingQueueBytesStorageServer    int                  `json:"limiting_queue_bytes_storage_server"`
+	LimitingQueueBytesStorageServer    int64                `json:"limiting_queue_bytes_storage_server"`
 	PerformanceLimitedBy               PerformanceLimitedBy `json:"performance_limited_by"`
 	ReleasedTransactionsPerSecond      float64              `json:"released_transactions_per_second"`
 	ThrottledTags                      ThrottledTags        `json:"throttled_tags"`
 	TransactionsPerSecondLimit         float64              `json:"transactions_per_second_limit"`
 	WorstDataLagStorageServer          Lag                  `json:"worst_data_lag_storage_server"`
 	WorstDurabilityLagStorageServer    Lag                  `json:"worst_durability_lag_storage_server"`
-	WorstQueueBytesLogServer           int                  `json:"worst_queue_bytes_log_server"`
-	WorstQueueBytesStorageServer       int                  `json:"worst_queue_bytes_storage_server"`
+	WorstQueueBytesLogServer           int64                `json:"worst_queue_bytes_log_server"`
+	WorstQueueBytesStorageServer       int64                `json:"worst_queue_bytes_storage_server"`
 }
 
 type PerformanceLimitedBy struct {
 	Description string `json:"description"`
 	Name        string `json:"name"`
-	ReasonId    int    `json:"reason_id"`
+	ReasonId    int64  `json:"reason_id"`
 }
 
 type ThrottledTags struct {
@@ -43,12 +43,12 @@ type ThrottledTags struct {
 }
 
 type AutoThrottledTags struct {
-	BusyRead        int `json:"busy_read"`
-	BusyWrite       int `json:"busy_write"`
-	Count           int `json:"count"`
-	RecommendedOnly int `json:"recommended_only"`
+	BusyRead        int64 `json:"busy_read"`
+	BusyWrite       int64 `json:"busy_write"`
+	Count           int64 `json:"count"`
+	RecommendedOnly int64 `json:"recommended_only"`
 }
 
 type ManualThrottledTags struct {
-	Count int `json:"count"`
+	Count int64 `json:"count"`
 }
